backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.22
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
langchain==0.3.27
langchain-google-genai==2.0.7
langchain-community==0.3.27
neo4j==5.21.0
supabase==2.5.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
httpx==0.27.0
